please check this patch!
#1. make sure the file is deleted if there is an error.
#2. move the ois.close() into the finally block to also ensure it is
closed on error.
#3. improve the logging messages
PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/tomcat@286189 13f79535-47bb-0310-9956-ffa450edef68

+++ container/tc4.x/trunk/catalina/src/share/org/apache/catalina/session/StandardManager.java
@@ -349,7 +349,7 @@ public class StandardManager
     public void load() throws ClassNotFoundException, IOException {
 
         if (debug >= 1)
-            log("Loading persisted sessions");
+            log("Start: Loading persisted sessions");
 
         // Initialize our internal data structures
         recycled.clear();
@@ -432,22 +432,23 @@ public class StandardManager
                     ois = null;
                 }
                 throw e;
+            } finally {
+                // Close the input stream
+                try {
+                    if (ois != null)
+                        ois.close();
+                } catch (IOException f) {
+                    // ignored
+                }
+
+                // Delete the persistent storage file
+                if (file != null && file.exists() )
+                    file.delete();
             }
         }
 
-        // Close the input stream
-        try {
-            ois.close();
-        } catch (IOException f) {
-            ;
-        }
-
-        // Delete the persistent storage file
-        file.delete();
-
         if (debug >= 1)
-            log("Loading complete");
-
+            log("Finish: Loading persisted sessions");
     }
